+++ actionview/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Introduce `config.action_view.dom_testing_assertions`
+
+    Adds support for `:rails_dom_testing` to support `Rails::Dom::Testing::Assertions` and `:none`.
+    Defaults to with `:rails_dom_testing`.
+
+    *Sean Doyle*
+
 *   Deprecate registering dependency trackers after application initialization.
 
     Calling `ActionView::DependencyTracker.register_tracker` after application initialization will